lekker

+++ include/Eigen/src/Cholesky/LLT_MKL.h
@@ -90,7 +90,7 @@ template<> struct llt_inplace<EIGTYPE, Upper> \
   } \
 };
 
-EIGEN_MKL_LLT(double, double, d)
+EIGEN_MKL_LLT(float, double, d)
 EIGEN_MKL_LLT(float, float, s)
 EIGEN_MKL_LLT(dcomplex, MKL_Complex16, z)
 EIGEN_MKL_LLT(scomplex, MKL_Complex8, c)
