+++ include/Eigen/src/LU/PartialPivLU_MKL.h
@@ -73,7 +73,7 @@ struct partial_lu_impl<EIGTYPE, StorageOrder, lapack_int> \
   } \
 };
 
-EIGEN_MKL_LU_PARTPIV(double, double, d)
+EIGEN_MKL_LU_PARTPIV(float, double, d)
 EIGEN_MKL_LU_PARTPIV(float, float, s)
 EIGEN_MKL_LU_PARTPIV(dcomplex, MKL_Complex16, z)
 EIGEN_MKL_LU_PARTPIV(scomplex, MKL_Complex8, c)
